refactor(pkg/operator): extracts types to a separate file
